requirements.txt
langchain==0.3.13
langchain_community==0.3.13
langchain_core==0.3.28
langchain_openai==0.3.0
PyPDF2==3.0.1
python-dotenv==1.0.1
streamlit==1.41.1
faiss-cpu==1.9.0.post1